Adding English Ending Mod
Adding the entirety of the English Ending Mod.

ModResources contains a sample of the files consumed by the mod. The contents would be included alongside the other mod's ModResources.

The mod contains 2 settings hidden in code. developerMode enables hotkeys that assist with debugging endings. blockTranslation blocks translations in case we need to extract the original text.